pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.5",
    "qrcode>=8.2",
]
